cloud function/requirements.txt
google-cloud-bigquery==2.10.0
pandas==1.2.3
pyarrow==3.0.0